+++ packages/terra-slide-panel/tests/wdio/slide-panel-spec.js
@@ -107,7 +107,7 @@ Terra.describeViewports('Slide panel', ['large'], () => {
       browser.pause(150);
 
       expect($('#test-toggle').isFocused()).toBeTruthy();
-      Terra.validates.element('toggle button focused', { selector: '#root' });
+      Terra.validates.element('toggle button focused mouse', { selector: '#root' });
     });
     it('Closes panel and focuses on toggle button with multiple buttons with keyboard controls', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/slide-panel/slide-panel-multiple-buttons-toggle');
@@ -149,7 +149,7 @@ Terra.describeViewports('Slide panel', ['large'], () => {
 
       expect($('#test-toggle').isFocused()).toBeTruthy();
 
-      Terra.validates.element('toggle button focused with multiple buttons', { selector: '#root' });
+      Terra.validates.element('toggle button focused with multiple buttons mouse', { selector: '#root' });
     });
   });
 
@@ -193,6 +193,34 @@ Terra.describeViewports('Slide panel', ['large'], () => {
 
       expect($('#close-panel-button').isFocused()).toBeFalsy();
       expect($('#toggle-panel-button').isFocused()).toBeFalsy();
+    });
+  });
+
+  describe('Non-focusable element used to disclose slide panel', () => {
+    it('does not focus SVG when it is the disclosing node', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/slide-panel/slide-panel-svg-toggle');
+      $('#test-click-svg').moveTo();
+      $('#test-click-svg').click();
+      $('#test-slide [aria-label="Panel content area"][aria-hidden="false"]').waitForExist();
+      browser.keys(['Tab']);
+      expect($('#focus-button').isFocused()).toBeTruthy();
+      browser.keys(['Enter']);
+
+      expect($('body').isFocused()).toBeTruthy();
+      expect($('#test-click-svg').isFocused()).toBeFalsy();
+    });
+
+    it('focuses main node when disclosing node has undefined focus', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/slide-panel/slide-panel-svg-toggle');
+      $('#test-p-click').moveTo();
+      $('#test-p-click').click();
+      $('#test-slide [aria-label="Panel content area"][aria-hidden="false"]').waitForExist();
+      browser.keys(['Tab']);
+      expect($('#focus-button').isFocused()).toBeTruthy();
+      browser.keys(['Enter']);
+
+      expect($('[aria-label="Main content area"]').isFocused()).toBeTruthy();
+      expect($('#test-click-svg').isFocused()).toBeFalsy();
     });
   });
 });