+++ packages/terra-tabs/src/common-tabs/_Tabs.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable class-methods-use-this */
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
@@ -5,6 +6,7 @@ import ThemeContext from 'terra-theme-context';
 import ResizeObserver from 'resize-observer-polyfill';
 import { v4 as uuid } from 'uuid';
 import IconAdd from 'terra-icon/lib/icon/IconAdd';
+import { DragDropContext, Droppable } from 'react-beautiful-dnd';
 import AddButton from './_AddButton';
 import MoreButton from './_MoreButton';
 import TabDropDown from './_TabDropDown';
@@ -90,6 +92,14 @@ const propTypes = {
    * It receives the updated tab data after a tab is closed.
    */
   onTabStateChange: PropTypes.func,
+  /**
+  * Whether or not the tab is draggable.
+  */
+  isDraggable: PropTypes.bool,
+  /**
+  * Callback function triggered when tab is drag and dropped .
+  */
+  onTabOrderChange: PropTypes.func,
 };
 
 let addTabId;
@@ -101,10 +111,6 @@ class Tabs extends React.Component {
     this.dropdownRef = React.createRef();
     this.moreButtonRef = React.createRef();
     this.addButtonRef = React.createRef();
-    this.state = {
-      tabData: this.props.tabData,
-    };
-
     this.setIsOpen = this.setIsOpen.bind(this);
     this.resetCache = this.resetCache.bind(this);
     this.handleResize = this.handleResize.bind(this);
@@ -117,7 +123,14 @@ class Tabs extends React.Component {
     this.wrapOnSelectHidden = this.wrapOnSelectHidden.bind(this);
     this.wrapOnClose = this.wrapOnClose.bind(this);
     this.wrapOnAddButton = this.wrapOnAddButton.bind(this);
+    this.positionDropDown = this.positionDropDown.bind(this);
+    this.handleDragEnd = this.handleDragEnd.bind(this);
+    this.handleDragStart = this.handleDragStart.bind(this);
+    this.reorder = this.reorder.bind(this);
     this.resetCache();
+    this.state = {
+      visibleTabData: this.props.tabData,
+    };
   }
 
   componentDidMount() {
@@ -136,19 +149,31 @@ class Tabs extends React.Component {
   }
 
   componentDidUpdate(prevProps) {
-    const newActiveKey = this.props.tabData.map(child => child.isSelected);
-    const prevActiveKey = prevProps.tabData.map(child => child.isSelected);
-    const isActiveKeyUpdate = JSON.stringify(newActiveKey) !== JSON.stringify(prevActiveKey);
+    const prevTab = prevProps.tabData.find((tab) => tab.isSelected === true);
+    const currTab = this.props.tabData.find((tab) => tab.isSelected === true);
+
+    // Allow dynamic addition of tabs.
+    if (this.state.visibleTabData.length !== this.props.tabData.length) {
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({ visibleTabData: this.props.tabData });
+    }
+
+    // Allow Active Styles to be applied when tab is selected.
+    if (prevTab.id !== currTab.id) {
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState(prevArray => {
+        const newArray = [...prevArray.visibleTabData];
+        const prevTabData = newArray.find(tab => tab.id === prevTab.id);
+        const currTabData = newArray.find(tab => tab.id === currTab.id);
+        prevTabData.isSelected = false;
+        currTabData.isSelected = true;
+        return { visibleTabData: newArray };
+      });
+    }
     if (this.isCalculating) {
       this.isCalculating = false;
       this.handleResize(this.contentWidth);
-      if (isActiveKeyUpdate) {
-        // eslint-disable-next-line react/no-did-update-set-state
-        this.setState({ tabData: this.props.tabData });
-      }
     } else if (this.props.tabData.length !== prevProps.tabData.length) {
-      // eslint-disable-next-line react/no-did-update-set-state
-      this.setState({ tabData: this.props.tabData });
       this.resetCache();
       this.forceUpdate();
     }
@@ -175,7 +200,7 @@ class Tabs extends React.Component {
     const availableWidth = width - moreButtonWidth - addtab;
 
     // Calculate hidden index
-    const tabCount = this.state.tabData.length;
+    const tabCount = this.state.visibleTabData.length;
     let newHideIndex = tabCount;
     let calcMinWidth = 0;
     let showMoreButton = false;
@@ -234,10 +259,46 @@ class Tabs extends React.Component {
     this.setIsOpen(false);
   }
 
+  handleDragStart() {
+    const tablist = document.querySelectorAll('[data-terra-drag-focus="true"]');
+    tablist.forEach((list) => {
+      list.setAttribute('data-terra-drag-focus', 'false');
+    });
+  }
+
+  handleDragEnd(result) {
+    const tablist = document.querySelectorAll('[data-terra-drag-focus="false"]');
+    tablist.forEach((list) => {
+      list.setAttribute('data-terra-drag-focus', 'true');
+    });
+    if (!result.destination) {
+      return;
+    }
+    this.setState((prevState) => {
+      const items = this.reorder(
+        prevState.visibleTabData,
+        result.source.index,
+        result.destination.index,
+      );
+      return { visibleTabData: items };
+    });
+    if (this.props.onTabOrderChange) {
+      this.props.onTabOrderChange(result);
+    }
+  }
+
   setIsOpen(value) {
     this.isOpen = value;
     this.forceUpdate();
   }
+
+  reorder = (list, startIndex, endIndex) => {
+    const result = Array.from(list);
+    const [removed] = result.splice(startIndex, 1);
+    result.splice(endIndex, 0, removed);
+
+    return result;
+  };
 
   resetCache() {
     this.animationFrameID = null;
@@ -277,11 +338,11 @@ class Tabs extends React.Component {
 
   wrapOnSelect(onSelect) {
     return (itemKey, metaData) => {
-      const updatedTabData = this.state.tabData.map((tab) => ({
+      const updatedTabData = this.state.visibleTabData.map((tab) => ({
         ...tab,
         isSelected: tab.itemKey === itemKey,
       }));
-      this.setState({ tabData: updatedTabData }, () => {
+      this.setState({ visibleTabData: updatedTabData }, () => {
         this.setIsOpen(false);
         onSelect(itemKey, metaData);
       });
@@ -291,11 +352,11 @@ class Tabs extends React.Component {
   wrapOnSelectHidden(onSelect) {
     return (itemKey, metaData) => {
       if (this.isOpen) {
-        const updatedTabData = this.state.tabData.map((tab) => ({
+        const updatedTabData = this.state.visibleTabData.map((tab) => ({
           ...tab,
           isSelected: tab.id === itemKey,
         }));
-        this.setState({ tabData: updatedTabData }, () => {
+        this.setState({ visibleTabData: updatedTabData }, () => {
           onSelect(itemKey, metaData);
           this.setIsOpen(!this.isOpen);
         });
@@ -303,22 +364,11 @@ class Tabs extends React.Component {
     };
   }
 
-  wrapOnAddButton() {
-    console.log('Sam Text', this.state.tabData);
-    console.log('hiddenStartIndex', this.hiddenStartIndex);
-    // const element = document.getElementById(this.state.tabData.id[this.state.tabData.id.length - 1]);
-    // element.focus();
-    if (this.props.onSelectAddButton) {
-      this.props.onSelectAddButton();
-    }
-    console.log('Sam Text', this.state.tabData);
-  }
-
   wrapOnClose(onClose) {
     return (itemKey, metaData, event) => {
       this.setIsOpen(false);
       let removedTabIndex = -1;
-      const updatedTabData = this.state.tabData
+      const updatedTabData = this.state.visibleTabData
         .map((tab, index) => {
           if (tab.itemKey === itemKey && tab.isSelected === true) {
             removedTabIndex = index;
@@ -341,7 +391,7 @@ class Tabs extends React.Component {
         }
       }
       this.props.onTabStateChange(updatedTabData, itemKey, event);
-      this.setState({ tabData: updatedTabData });
+      this.setState({ visibleTabData: updatedTabData });
       onClose(itemKey, metaData);
     };
   }
@@ -349,10 +399,10 @@ class Tabs extends React.Component {
   render() {
     addTabId = uuid();
     const {
-      ariaLabel, variant, onChange, onSelectAddButton, ariaLabelAddTab,
+      ariaLabel, variant, onChange, onSelectAddButton, ariaLabelAddTab, isDraggable,
     } = this.props;
     const theme = this.context;
-    const enabledTabs = this.state.tabData.filter(tab => !tab.isDisabled);
+    const enabledTabs = this.state.visibleTabData.filter(tab => !tab.isDisabled);
     const ids = enabledTabs.map(tab => tab.id);
     const hiddenIds = [];
     const visibleTabs = [];
@@ -362,11 +412,11 @@ class Tabs extends React.Component {
     let isHiddenSelected = false;
 
     let enabledTabsIndex = -1;
-    this.state.tabData.forEach((tab, index) => {
+    this.state.visibleTabData.forEach((tab, index) => {
       if (!tab.isDisabled) {
         enabledTabsIndex += 1;
       }
-      if (enabledTabsIndex < this.hiddenStartIndex || this.hiddenStartIndex < 0) {
+      if (index < this.hiddenStartIndex || this.hiddenStartIndex < 0) {
         visibleTabs.push(
           <Tab
             {...tab}
@@ -376,7 +426,7 @@ class Tabs extends React.Component {
             icon={tab.icon}
             customDisplay={tab.customDisplay}
             onSelect={this.wrapOnSelect(tab.onSelect)}
-            zIndex={tab.isSelected ? this.state.tabData.length : this.state.tabData.length - index}
+            zIndex={tab.isSelected ? this.state.visibleTabData.length : this.state.visibleTabData.length - index}
             isIconOnly={tab.isIconOnly}
             variant={variant}
             onChange={onChange}
@@ -384,6 +434,7 @@ class Tabs extends React.Component {
             hiddenStartIndex={this.hiddenStartIndex}
             showIcon={tab.showIcon}
             onClosingTab={this.wrapOnClose(tab.onClose)}
+            isDraggable={isDraggable}
           />,
         );
       } else {
@@ -400,6 +451,7 @@ class Tabs extends React.Component {
             icon={tab.icon}
             showIcon={tab.showIcon}
             onClosingTab={this.wrapOnClose(tab.onClose)}
+            isDisabled={tab.isDisabled}
           />,
         );
         hiddenIds.push(tab.id);
@@ -407,14 +459,14 @@ class Tabs extends React.Component {
         if (tab.isSelected) {
           isHiddenSelected = true;
         }
-        if (index === this.state.tabData.length - 1 && onSelectAddButton) {
+        if (index === this.state.visibleTabData.length - 1 && onSelectAddButton) {
           hiddenTabs.push(
             <HiddenTab
               id={addTabId}
               data-focus-styles-enabled
               itemKey={addTabId}
               label={ariaLabelAddTab}
-              index={this.state.tabData.length}
+              index={this.state.visibleTabData.length}
               showIcon
               icon={<IconAdd a11yLabel={ariaLabelAddTab} />}
               tabIds={moreIds}
@@ -440,6 +492,58 @@ class Tabs extends React.Component {
       };
     }
     const commonTabsClassNames = cx('tab-container', theme.className);
+    window['__react-beautiful-dnd-disable-dev-warnings'] = true;
+
+    if (isDraggable) {
+      return (
+        <DragDropContext onDragStart={this.handleDragStart} onDragEnd={this.handleDragEnd}>
+          <Droppable className={commonTabsClassNames} droppableId="tab-list" direction="horizontal">
+            {(provided) => (
+              <div
+                {...attrs}
+                {...provided.droppableProps}
+                ref={(el) => {
+                  provided.innerRef(el);
+                  this.containerRef.current = el; // Store the reference to the container element
+                }}
+                className={commonTabsClassNames}
+                role="tablist"
+                aria-label={ariaLabel}
+                aria-orientation="horizontal"
+                aria-owns={hiddenIds.join(' ')}
+                data-terra-drag-focus
+              >
+                {visibleTabs}
+                {provided.placeholder}
+                {this.showMoreButton ? (
+                  <MoreButton
+                    isOpen={this.isOpen}
+                    hiddenIndex={this.hiddenStartIndex}
+                    isActive={isHiddenSelected}
+                    zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
+                    onBlur={this.handleMoreButtonBlur}
+                    onSelect={this.handleMoreButtonSelect}
+                    refCallback={node => { this.moreButtonRef.current = node; }}
+                    tabIds={ids}
+                    variant={variant}
+                  />
+                ) : undefined}
+                <TabDropDown
+                  onFocus={this.handleHiddenFocus}
+                  onBlur={this.handleHiddenBlur}
+                  isOpen={this.isOpen}
+                  onRequestClose={this.handleOutsideClick}
+                  refCallback={node => { this.dropdownRef.current = node; }}
+                >
+                  {hiddenTabs}
+                </TabDropDown>
+              </div>
+            )}
+          </Droppable>
+        </DragDropContext>
+      );
+    }
+
     const commonTabsContainerClassNames = cx('container', theme.className);
     const commonDivClassNames = cx('divcontainer', theme.className);
 
@@ -447,8 +551,8 @@ class Tabs extends React.Component {
       <div className={commonTabsContainerClassNames}>
         <div
           {...attrs}
-          className={commonTabsClassNames}
           ref={this.containerRef}
+          className={commonTabsClassNames}
           role="tablist"
           aria-label={ariaLabel}
           aria-orientation="horizontal"
@@ -460,7 +564,7 @@ class Tabs extends React.Component {
               isOpen={this.isOpen}
               hiddenIndex={this.hiddenStartIndex}
               isActive={isHiddenSelected}
-              zIndex={this.state.tabData.length - this.hiddenStartIndex}
+              zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
               onBlur={this.handleMoreButtonBlur}
               onSelect={this.handleMoreButtonSelect}
               refCallback={node => { this.moreButtonRef.current = node; }}
@@ -484,12 +588,11 @@ class Tabs extends React.Component {
             id={addTabId}
             addAriaLabel={ariaLabelAddTab}
             index={enabledTabsIndex + 1}
-            onSelect={this.wrapOnAddButton}
+            onSelect={onSelectAddButton}
             tabIds={moreIds}
             isSelected={false}
           />
           )}
-
         </div>
       </div>
     );