+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 3.89.0 - (August 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.88.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.87.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.86.0 - (July 19, 2023)
 
 * Changed