+++ packages/terra-form-validation/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 1.84.0 - (August 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.83.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.82.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.81.0 - (July 19, 2023)
 
 * Changed