+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 6.74.0 - (August 2, 2023)
+
+* Changed
+  * Locked `uuid` dependency to `3.4.0` for consistency across Terra packages.
+
+## 6.73.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 6.72.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.71.0 - (July 19, 2023)
 
 * Changed