+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.33.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.32.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.31.0 - (July 19, 2023)
 
 * Changed