+++ packages/terra-pills/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 1.11.0 - (August 2, 2023)
+
+* Changed
+  * Locked `uuid` dependency to `3.4.0` for consistency across Terra packages.
+
+## 1.10.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.9.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.8.0 - (July 19, 2023)
 
 * Changed