+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 6.70.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 6.69.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.68.0 - (July 19, 2023)
 
 * Changed