+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 5.76.0 - (August 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 5.75.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 5.74.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 5.73.0 - (July 19, 2023)
 
 * Changed