+++ packages/terra-slide-panel/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+## 3.45.0 - (August 2, 2023)
+
+* Changed
+  * Locked `uuid` dependency to `3.4.0` for consistency across Terra packages.
+
+## 3.44.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+* Added
+  * Added check for focus on the disclosing node before calling focus.
+
 ## 3.43.0 - (July 19, 2023)
 
 * Changed